Change train method in Classifier to accept double[] instead of int[]

+++ Part2/Learning/src/main/java/classification/Classifier.java
@@ -8,7 +8,7 @@ public abstract class Classifier {
         throw new UnsupportedOperationException();
     }
     
-    public EmailClass classify(int[] vector) {
+    public EmailClass classify(double[] vector) {
         throw new UnsupportedOperationException();
     }
 }
